refactor(agent): move durable agents into durable module directory
