+++ packages/agent-manager/src/durable/ClaudeCliProbe.ts
